+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Simplified Monopoly™ turns prediction".
 - Tests for "Regular polygons".
 - Tests for "Horse-hyperracing hyperduals".
+- Tests for "3×N tiling".
 
 ## [1.13.0] - 2022-09-30
 ### Added